Pixel pen rotation bug

+++ src/main/java/chalmers/pimp/model/canvas/CanvasImpl.java
@@ -85,7 +85,7 @@ final class CanvasImpl implements ICanvas {
   }
 
   @Override
-  public void setActiveLayerRotation(double alpha) {
+  public void setActiveLayerRotation(int alpha) {
     layerManager.rotateActiveLayer(alpha);
   }
