+++ src/main/java/chalmers/pimp/model/canvas/ICanvas.java
@@ -102,7 +102,7 @@ public interface ICanvas extends IMementoTarget<CanvasMemento>, ICopiable<ICanva
    *
    * @param alpha the rotation in degrees.
    */
-  void setActiveLayerRotation(double alpha);
+  void setActiveLayerRotation(int alpha);
 
   /**
    * Sets the visibility property of the layer associated with the supplied depth index. This method
