+++ src/main/java/chalmers/pimp/model/canvas/layer/Rectangle.java
@@ -84,7 +84,7 @@ final class Rectangle implements ILayer, IColorable {
   }
 
   @Override
-  public void setRotation(double rotation) {
+  public void setRotation(int rotation) {
     layerDelegate.setRotationDegrees(rotation);
   }
 
@@ -129,7 +129,7 @@ final class Rectangle implements ILayer, IColorable {
   }
 
   @Override
-  public double getRotation() {
+  public int getRotation() {
     return layerDelegate.getRotationDegrees();
   }
